ASoC: wm_adsp: Move DSP Rate controls into the codec
The rate controls are codec-specific, it's not possible to
generically say what the range or the meaning of each control
is (or even if they exist at all) - that depends on the
particular codec.

This is currently being handled for Arizona codecs by putting
an Arizona-specific table of controls inside the wm_adsp driver.
This creates a dependency between wm_adsp and arizona.c, and is an
awkward solution if the ADSP is used in another family of codecs

Fix this by moving the Arizona-specific rate controls into the
Arizona codec drivers.

Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 336d0442b9f03a0389baaf09e39880fdd1873cc3)

Also added support for ADSP2v2 rate controls on Clearwater, Marley, Moon.

Change-Id: Ib9a548247dd176618bc247362e504e2b00b57078
Signed-off-by: Richard Fitzgerald <rf@opensource.wolfsonmicro.com>